requirements.txt
asttokens==2.4.1
Brotli==1.0.9
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fasttext-wheel==0.9.2
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.5.0
ftfy==6.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.2
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
libretranslatepy==2.1.1
llvmlite==0.42.0
lxml==5.2.2
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pybind11==2.12.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchtext==0.17.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
translate==3.6.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
